bad behavier again...
I learned a lot about c++, and I tried to apply them to this game, and I
did it. :-)
by now, everything seems fine. In this commit, I clean the old project
file, and refactored almost everything. (because they were totally a
mess...) Now we have a background, a hero which has a EVIL weapon that
allways hurt himself, I tried my best to make everything flexible, and I
hope it is....  that all.